requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
jinja2==3.1.4
python-multipart==0.0.9
scikit-learn==1.4.2
pandas==2.2.2
python-dotenv==1.0.1
